chore: prepare VIP-656 for proposal
Renumber vip-999 -> vip-656 (on-chain proposalCount is 655), retitle from
the Notion VIP Voting doc, and restore minBorrowCap to 500,000 USDT, which
94ed95f had set and 69e2ea5 reverted to 10.

The description now lists all 16 actions in calldata order (the pause /
unpause grants added in 69e2ea5 were missing) and cites the redeployed
vceBTC at 0xAB7D138c6e6fF1bfD3ac871d0dB08f9442Ce927F.
