+++ .gitignore
@@ -40,3 +40,6 @@ scripts/data/safePauseTxMetadata*.json
 
 .cache
 .vscode/settings.json
+
+safe-propose.json
+safe-vote.json
