+++ simulations/vip-656/abi/ResilientOracle.json
